+++ src/mol-model-formats/structure/basic/coarse.ts
@@ -16,6 +16,7 @@ import { ElementIndex, ChainIndex } from '../../../mol-model/structure/model/ind
 import { getCoarseRanges } from '../../../mol-model/structure/model/properties/utils/coarse-ranges';
 import { IhmSphereObjSite, IhmGaussianObjSite, AtomSite, BasicSchema } from './schema';
 import { Model } from '../../../mol-model/structure';
+import { getCoarseIndex } from '../../../mol-model/structure/model/properties/utils/coarse-index';
 
 export interface CoarseData {
     model_id: number,
@@ -50,6 +51,7 @@ export function getCoarse(data: CoarseData, chemicalComponentMap: Model['propert
             isDefined: true,
             spheres: { ...sphereData, ...sphereKeys, ...sphereRanges },
             gaussians: { ...gaussianData, ...gaussianKeys, ...gaussianRanges },
+            index: getCoarseIndex({ spheres: sphereData, gaussians: gaussianData })
         },
         conformation: {
             id: UUID.create22(),